+++ AirBean/src/GlobalCSS/_index.scss
@@ -1,1 +1,7 @@
 @forward "./Variables.scss";
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
