+++ AirBean/src/components/CartOverlay/CartOverlay.scss
@@ -1,6 +1,6 @@
 
 .cart-overlay-background {
-  position: fixed;
+  position: absolute;
   top: 0; 
   left: 0;
   width: 100vw;
