+++ AirBean/src/index.css
@@ -1,1 +1,1 @@
-@import url("./Components/Navbar/Navbar.scss");
+@import "./GlobalCSS/_index.scss"
