+++ AirBean/src/pages/Status/Status.scss
